+++ R/block-server.R
@@ -228,7 +228,8 @@ reorder_dots_observer <- function(data, sess) {
           reorder_rv(data[["...args"]], arg_names[ind])
         }
       },
-      domain = sess
+      domain = sess,
+      label = otel_lbl("reorder_dots")
     )
   }
 }
@@ -255,7 +256,8 @@ validate_block_observer <- function(id, x, dat, res, rv, cond, sess) {
           session = sess
         )
       },
-      domain = sess
+      domain = sess,
+      label = otel_lbl("validate_block")
     )
   }
 }
@@ -312,7 +314,8 @@ state_check_observer <- function(id, x, dat, res, state, rv, cond, sess) {
         rv$state_set <- TRUE
       }
     },
-    domain = sess
+    domain = sess,
+    label = otel_lbl("check_state")
   )
 }
 
@@ -333,7 +336,8 @@ data_eval_observer <- function(id, x, dat, gate, res, state, lang, rv, cond,
 
       res(out)
     },
-    domain = sess
+    domain = sess,
+    label = otel_lbl("eval_data")
   )
 }
 
@@ -379,7 +383,8 @@ output_render_observer <- function(x, res, cond, sess) {
         session = sess
       )
     },
-    domain = sess
+    domain = sess,
+    label = otel_lbl("render_output")
   )
 }
 
@@ -412,7 +417,8 @@ block_cond_observer <- function(blk, cond, sess) {
       } else if (any(lengths(cur_cnds) > 0L)) {
         cond$block <- empty_block_condition()[names(new_cnds)]
       }
-    }
+    },
+    label = otel_lbl("update_block_cond")
   )
 }
 
@@ -511,7 +517,8 @@ check_expr_val <- function(val, x) {
               }
             }
           },
-          once = TRUE
+          once = TRUE,
+          label = otel_lbl(paste0("check_cond_", cnd))
         )
       },
       val[["cond"]]