+++ R/plugin-block.R
@@ -53,7 +53,8 @@ edit_block_server <- function(id, block_id, board, update, ...) {
           "block_name_in",
           "Block name",
           cur_name()
-        )
+        ),
+        label = otel_lbl("sync_block_name_input")
       )
 
       observeEvent(
@@ -66,7 +67,8 @@ edit_block_server <- function(id, block_id, board, update, ...) {
             new_val <- as_blocks(set_names(list(new_val), block_id))
             update(list(blocks = list(mod = new_val)))
           }
-        }
+        },
+        label = otel_lbl("update_block_name")
       )
 
       output$block_summary <- renderText(
@@ -78,7 +80,8 @@ edit_block_server <- function(id, block_id, board, update, ...) {
 
       observeEvent(
         input$rm_block,
-        update(list(blocks = list(rm = block_id)))
+        update(list(blocks = list(rm = block_id))),
+        label = otel_lbl("rm_block")
       )
 
       position <- NULL
@@ -106,7 +109,8 @@ edit_block_server <- function(id, block_id, board, update, ...) {
                 session = session
               )
             }
-          }
+          },
+          label = otel_lbl("add_block_before")
         )
       }
 
@@ -119,7 +123,8 @@ edit_block_server <- function(id, block_id, board, update, ...) {
             position,
             stack_checkbox = is_block_in_stacks(block_id, board$board)
           )
-        }
+        },
+        label = otel_lbl("add_block_after")
       )
 
       observeEvent(
@@ -146,7 +151,8 @@ edit_block_server <- function(id, block_id, board, update, ...) {
               available_block_inputs(sel)
             )
           }
-        }
+        },
+        label = otel_lbl("update_insert_input_choices")
       )
 
       observeEvent(
@@ -224,12 +230,14 @@ edit_block_server <- function(id, block_id, board, update, ...) {
           removeModal()
 
           position <<- NULL
-        }
+        },
+        label = otel_lbl("confirm_insert_block")
       )
 
       observeEvent(
         input$cancel_insert,
-        removeModal()
+        removeModal(),
+        label = otel_lbl("cancel_insert_block")
       )
 
       NULL